internal/manifest: move TableMetadata to a separate file
TableMetadata and related types have grown in code volume. Move them to a file
separation from Version.
